# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SYR SYP Currency Exchange API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: exchange
    description: Currency exchange rates
  - name: gold
    description: Gold prices
  - name: news
    description: Currency and gold news
  - name: profile
    description: User profile management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /exchange/rates:
    get:
      operationId: getExchangeRates
      tags: [exchange]
      summary: Get all supported currency exchange rates
      responses:
        "200":
          description: Exchange rates list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRatesResponse"

  /exchange/convert:
    get:
      operationId: convertCurrency
      tags: [exchange]
      summary: Convert currency amount
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
        - name: to
          in: query
          required: true
          schema:
            type: string
        - name: amount
          in: query
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Conversion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversionResult"

  /gold/prices:
    get:
      operationId: getGoldPrices
      tags: [gold]
      summary: Get gold prices by karat
      responses:
        "200":
          description: Gold prices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoldPricesResponse"

  /news:
    get:
      operationId: getNews
      tags: [news]
      summary: Get latest currency and gold news
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Search query for news
      responses:
        "200":
          description: News list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsItem"

  /metals:
    get:
      operationId: getMetalPrices
      tags: [exchange]
      summary: Get all precious metal prices
      responses:
        "200":
          description: Metal prices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetalPricesResponse"

  /alerts:
    get:
      operationId: getAlerts
      tags: [alerts]
      summary: Get user price alerts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of price alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceAlert"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createAlert
      tags: [alerts]
      summary: Create a new price alert
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlertBody"
      responses:
        "201":
          description: Created alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceAlert"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /alerts/{id}:
    put:
      operationId: updateAlert
      tags: [alerts]
      summary: Update a price alert
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlertBody"
      responses:
        "200":
          description: Updated alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceAlert"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAlert
      tags: [alerts]
      summary: Delete a price alert
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /alerts/check:
    post:
      operationId: checkAlerts
      tags: [alerts]
      summary: Check alerts against current rates and return triggered ones
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckAlertsBody"
      responses:
        "200":
          description: Triggered alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceAlert"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ExchangeRatesResponse:
      type: object
      properties:
        base:
          type: string
        rates:
          type: object
          additionalProperties:
            type: number
        usd_to_syp:
          type: number
        try_to_syp:
          type: number
        is_manual_rate:
          type: boolean
        timestamp:
          type: string
      required:
        - base
        - rates
        - usd_to_syp
        - try_to_syp
        - is_manual_rate
        - timestamp

    ConversionResult:
      type: object
      properties:
        from:
          type: string
        to:
          type: string
        amount:
          type: number
        result:
          type: number
        rate:
          type: number
        timestamp:
          type: string
      required:
        - from
        - to
        - amount
        - result
        - rate
        - timestamp

    GoldPricesResponse:
      type: object
      properties:
        pricePerGramUSD:
          type: number
        pricePerGramSYP:
          type: number
        karats:
          type: array
          items:
            $ref: "#/components/schemas/GoldKarat"
        timestamp:
          type: string
      required:
        - pricePerGramUSD
        - pricePerGramSYP
        - karats
        - timestamp

    GoldKarat:
      type: object
      properties:
        karat:
          type: integer
        purity:
          type: number
        pricePerGramUSD:
          type: number
        pricePerGramSYP:
          type: number
      required:
        - karat
        - purity
        - pricePerGramUSD
        - pricePerGramSYP

    MetalPricesResponse:
      type: object
      properties:
        metals:
          type: array
          items:
            $ref: "#/components/schemas/MetalPrice"
        timestamp:
          type: string
      required:
        - metals
        - timestamp

    MetalPrice:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        symbol:
          type: string
        priceUSD:
          type: number
        priceSYP:
          type: number
        unit:
          type: string
      required:
        - name
        - nameAr
        - symbol
        - priceUSD
        - priceSYP
        - unit

    NewsItem:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        summary:
          type: string
        source:
          type: string
        publishedAt:
          type: string
        category:
          type: string
      required:
        - id
        - title
        - summary
        - source
        - publishedAt
        - category

    UserProfile:
      type: object
      properties:
        clerkId:
          type: string
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        fatherName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        governorate:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        profilePhoto:
          type: ["string", "null"]
        loginProvider:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        lastSeenAt:
          type: ["string", "null"]
        isBanned:
          type: boolean
        bannedReason:
          type: ["string", "null"]
        bannedUntil:
          type: ["string", "null"]
        profileCompleted:
          type: boolean
      required:
        - clerkId
        - email
        - profileCompleted
        - isBanned

    UpdateProfileBody:
      type: object
      properties:
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        fatherName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        governorate:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        profilePhoto:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]

    PriceAlert:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        code:
          type: string
        nameAr:
          type: ["string", "null"]
        type:
          type: string
          enum: [buy, sell]
        targetPrice:
          type: number
        isTriggered:
          type: boolean
        triggeredAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - clerkId
        - code
        - type
        - targetPrice
        - isTriggered
        - createdAt
        - updatedAt

    CreateAlertBody:
      type: object
      properties:
        code:
          type: string
        nameAr:
          type: string
        type:
          type: string
          enum: [buy, sell]
        targetPrice:
          type: number
      required:
        - code
        - type
        - targetPrice

    UpdateAlertBody:
      type: object
      properties:
        type:
          type: string
          enum: [buy, sell]
        targetPrice:
          type: number
        nameAr:
          type: string

    CheckAlertsBody:
      type: object
      properties:
        rates:
          type: object
          additionalProperties:
            type: number
      required:
        - rates

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error