# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SYR SYP Currency Exchange API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Get all supported currency exchange rates
 */
export const GetExchangeRatesResponse = zod.object({
  "base": zod.string(),
  "rates": zod.record(zod.string(), zod.number()),
  "usd_to_syp": zod.number(),
  "try_to_syp": zod.number(),
  "is_manual_rate": zod.boolean(),
  "timestamp": zod.string()
})


/**
 * @summary Convert currency amount
 */
export const ConvertCurrencyQueryParams = zod.object({
  "from": zod.coerce.string(),
  "to": zod.coerce.string(),
  "amount": zod.coerce.number()
})

export const ConvertCurrencyResponse = zod.object({
  "from": zod.string(),
  "to": zod.string(),
  "amount": zod.number(),
  "result": zod.number(),
  "rate": zod.number(),
  "timestamp": zod.string()
})


/**
 * @summary Get gold prices by karat
 */
export const GetGoldPricesResponse = zod.object({
  "pricePerGramUSD": zod.number(),
  "pricePerGramSYP": zod.number(),
  "karats": zod.array(zod.object({
  "karat": zod.number(),
  "purity": zod.number(),
  "pricePerGramUSD": zod.number(),
  "pricePerGramSYP": zod.number()
})),
  "timestamp": zod.string()
})


/**
 * @summary Get latest currency and gold news
 */
export const GetNewsQueryParams = zod.object({
  "q": zod.coerce.string().optional().describe('Search query for news')
})

export const GetNewsResponseItem = zod.object({
  "id": zod.string(),
  "title": zod.string(),
  "summary": zod.string(),
  "source": zod.string(),
  "publishedAt": zod.string(),
  "category": zod.string()
})
export const GetNewsResponse = zod.array(GetNewsResponseItem)


/**
 * @summary Get all precious metal prices
 */
export const GetMetalPricesResponse = zod.object({
  "metals": zod.array(zod.object({
  "name": zod.string(),
  "nameAr": zod.string(),
  "symbol": zod.string(),
  "priceUSD": zod.number(),
  "priceSYP": zod.number(),
  "unit": zod.string()
})),
  "timestamp": zod.string()
})


/**
 * @summary Get user price alerts
 */
export const GetAlertsResponseItem = zod.object({
  "id": zod.number(),
  "clerkId": zod.string(),
  "code": zod.string(),
  "nameAr": zod.string().nullish(),
  "type": zod.enum(['buy', 'sell']),
  "targetPrice": zod.number(),
  "isTriggered": zod.boolean(),
  "triggeredAt": zod.string().nullish(),
  "createdAt": zod.string(),
  "updatedAt": zod.string()
})
export const GetAlertsResponse = zod.array(GetAlertsResponseItem)


/**
 * @summary Create a new price alert
 */
export const CreateAlertBody = zod.object({
  "code": zod.string(),
  "nameAr": zod.string().optional(),
  "type": zod.enum(['buy', 'sell']),
  "targetPrice": zod.number()
})


/**
 * @summary Update a price alert
 */
export const UpdateAlertParams = zod.object({
  "id": zod.coerce.number()
})

export const UpdateAlertBody = zod.object({
  "type": zod.enum(['buy', 'sell']).optional(),
  "targetPrice": zod.number().optional(),
  "nameAr": zod.string().optional()
})

export const UpdateAlertResponse = zod.object({
  "id": zod.number(),
  "clerkId": zod.string(),
  "code": zod.string(),
  "nameAr": zod.string().nullish(),
  "type": zod.enum(['buy', 'sell']),
  "targetPrice": zod.number(),
  "isTriggered": zod.boolean(),
  "triggeredAt": zod.string().nullish(),
  "createdAt": zod.string(),
  "updatedAt": zod.string()
})


/**
 * @summary Delete a price alert
 */
export const DeleteAlertParams = zod.object({
  "id": zod.coerce.number()
})

export const DeleteAlertResponse = zod.object({
  "success": zod.boolean()
})


/**
 * @summary Check alerts against current rates and return triggered ones
 */
export const CheckAlertsBody = zod.object({
  "rates": zod.record(zod.string(), zod.number())
})

export const CheckAlertsResponseItem = zod.object({
  "id": zod.number(),
  "clerkId": zod.string(),
  "code": zod.string(),
  "nameAr": zod.string().nullish(),
  "type": zod.enum(['buy', 'sell']),
  "targetPrice": zod.number(),
  "isTriggered": zod.boolean(),
  "triggeredAt": zod.string().nullish(),
  "createdAt": zod.string(),
  "updatedAt": zod.string()
})
export const CheckAlertsResponse = zod.array(CheckAlertsResponseItem)


/**
 * @summary Get user profile
 */
export const GetProfileResponse = zod.object({
  "clerkId": zod.string(),
  "email": zod.string(),
  "firstName": zod.string().nullish(),
  "lastName": zod.string().nullish(),
  "fatherName": zod.string().nullish(),
  "phone": zod.string().nullish(),
  "birthDate": zod.string().nullish(),
  "gender": zod.string().nullish(),
  "governorate": zod.string().nullish(),
  "city": zod.string().nullish(),
  "address": zod.string().nullish(),
  "profilePhoto": zod.string().nullish(),
  "loginProvider": zod.string().nullish(),
  "latitude": zod.number().nullish(),
  "longitude": zod.number().nullish(),
  "lastSeenAt": zod.string().nullish(),
  "isBanned": zod.boolean(),
  "bannedReason": zod.string().nullish(),
  "bannedUntil": zod.string().nullish(),
  "profileCompleted": zod.boolean()
})


/**
 * @summary Update user profile
 */
export const UpdateProfileBody = zod.object({
  "firstName": zod.string().nullish(),
  "lastName": zod.string().nullish(),
  "fatherName": zod.string().nullish(),
  "phone": zod.string().nullish(),
  "birthDate": zod.string().nullish(),
  "gender": zod.string().nullish(),
  "governorate": zod.string().nullish(),
  "city": zod.string().nullish(),
  "address": zod.string().nullish(),
  "profilePhoto": zod.string().nullish(),
  "latitude": zod.number().nullish(),
  "longitude": zod.number().nullish()
})

export const UpdateProfileResponse = zod.object({
  "clerkId": zod.string(),
  "email": zod.string(),
  "firstName": zod.string().nullish(),
  "lastName": zod.string().nullish(),
  "fatherName": zod.string().nullish(),
  "phone": zod.string().nullish(),
  "birthDate": zod.string().nullish(),
  "gender": zod.string().nullish(),
  "governorate": zod.string().nullish(),
  "city": zod.string().nullish(),
  "address": zod.string().nullish(),
  "profilePhoto": zod.string().nullish(),
  "loginProvider": zod.string().nullish(),
  "latitude": zod.number().nullish(),
  "longitude": zod.number().nullish(),
  "lastSeenAt": zod.string().nullish(),
  "isBanned": zod.boolean(),
  "bannedReason": zod.string().nullish(),
  "bannedUntil": zod.string().nullish(),
  "profileCompleted": zod.boolean()
})